AG-10003 - Merge Nx plugins.
