--- plugins/example-generator/README.md
@@ -1,11 +1,0 @@
-# example-generator
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build example-generator` to build the library.
-
-## Running unit tests
-
-Run `nx test example-generator` to execute the unit tests via [Jest](https://jestjs.io).
