--- plugins/example-generator/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "example-generator",
-  "version": "0.0.1",
-  "dependencies": {
-    "tslib": "^2.3.0"
-  },
-  "type": "module",
-  "main": "./src/index.ts",
-  "typings": "./src/index.d.ts"
-}
