--- plugins/example-generator/project.json
@@ -1,62 +1,0 @@
-{
-  "name": "example-generator",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "plugins/example-generator/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nx/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/plugins/example-generator",
-        "main": "plugins/example-generator/src/index.ts",
-        "tsConfig": "plugins/example-generator/tsconfig.lib.json",
-        "assets": [
-          "plugins/example-generator/*.md",
-          {
-            "input": "./plugins/example-generator/src",
-            "glob": "**/!(*.ts)",
-            "output": "./src"
-          },
-          {
-            "input": "./plugins/example-generator/src",
-            "glob": "**/*.d.ts",
-            "output": "./src"
-          },
-          {
-            "input": "./plugins/example-generator",
-            "glob": "generators.json",
-            "output": "."
-          },
-          {
-            "input": "./plugins/example-generator",
-            "glob": "executors.json",
-            "output": "."
-          }
-        ]
-      }
-    },
-    "lint": {
-      "executor": "@nx/eslint:lint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["plugins/example-generator/**/*.ts", "plugins/example-generator/package.json"]
-      }
-    },
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "plugins/example-generator/jest.config.ts",
-        "passWithNoTests": true
-      },
-      "configurations": {
-        "ci": {
-          "ci": true,
-          "codeCoverage": true
-        }
-      }
-    }
-  },
-  "tags": []
-}
